requirements.txt
pycryptodome>=3.18.0
flask>=2.3.3
numpy>=1.24.0
matplotlib>=3.7.0
